+++ rs/src/lib.rs
@@ -2,6 +2,7 @@ extern crate swf_fixed;
 extern crate swf_tree;
 
 pub mod basic_data_types;
+pub mod button;
 pub mod bit_count;
 pub mod display;
 pub mod gradient;
@@ -19,16 +20,15 @@ mod tests {
   use std::path::Path;
 
   use swf_parser;
-  use swf_tree::{CompressionMethod, Movie, Tag};
+  use swf_tree::{ColorTransformWithAlpha, CompressionMethod, Header, Movie, SwfSignature, Tag};
   use swf_tree::Matrix;
   use swf_tree::Rect;
 
   use ::test_generator::test_expand_paths;
 
-  use crate::basic_data_types::emit_leb128_u32;
-  use crate::basic_data_types::emit_matrix;
-  use crate::basic_data_types::emit_rect;
-  use crate::movie::emit_movie;
+  use crate::basic_data_types::{emit_color_transform_with_alpha, emit_leb128_u32, emit_matrix, emit_rect};
+  use crate::movie::{emit_header, emit_movie, emit_swf_signature};
+  use crate::primitives::emit_le_f16;
   use crate::tags::emit_tag;
 
   test_expand_paths! { test_emit_movie; "../tests/movies/*/" }
@@ -138,9 +138,17 @@ mod tests {
     }
   }
 
+  test_various_ref_emitter_impl!(test_emit_color_transform_with_alpha, "../tests/various/color-transform-with-alpha/*/", emit_color_transform_with_alpha, ColorTransformWithAlpha);
+
+  test_various_copy_emitter_impl!(test_emit_le_f16, "../tests/various/float16-le/*/", emit_le_f16, f32);
+
+  test_various_ref_emitter_impl!(test_emit_header, "../tests/various/header/*/", emit_header, Header);
+
   test_various_ref_emitter_impl!(test_emit_matrix, "../tests/various/matrix/*/", emit_matrix, Matrix);
 
   test_various_ref_emitter_impl!(test_emit_rect, "../tests/various/rect/*/", emit_rect, Rect);
+
+  test_various_ref_emitter_impl!(test_emit_swf_signature, "../tests/various/swf-signature/*/", emit_swf_signature, SwfSignature);
 
   test_various_copy_emitter_impl!(test_emit_leb128_u32, "../tests/various/uint32-leb128/*/", emit_leb128_u32, u32);
 }